Changed tables, worked fine...BUT A PROBLEM!!!
Everything was fine until I switched to a bigger dataset. I should have
known. I should have seen this coming.
Anyway. the graph is still drawn very nicely. All the values are there
with thinner bars and that´s very cool.
However, the ticks on the axis don´t display correctly. They appear to
be all n the sides. I have to check what´s up with this, but I haven´t
found it yet.
But on the bright side, yay for displaying bigger datasets in the same
space.