Graph with interpolation
For the sake of testing the zoom, I´ve chosen to turn off the other two
graphs (the high and the volume) so that I can work only on the Close
line. using .curve(d3.curveLinear) the graph now interpolates between
points, making it look much more appealing and way less square.